+++ lib/active_admin/form_builder.rb
@@ -30,11 +30,6 @@ module ActiveAdmin
       form_buffers.last << template.content_tag(:li, (template.link_to I18n.t('active_admin.cancel'), url, html_options), li_attributes)
     end
 
-    def commit_action_with_cancel_link
-      content = action(:submit)
-      content << cancel_link
-    end
-
     def actions(*args, &block)
       content = with_new_form_buffer do
         block_given? ? super : super { commit_action_with_cancel_link }
@@ -59,17 +54,26 @@ module ActiveAdmin
 
       # Add Delete Links
       form_block = proc do |has_many_form|
-        block.call(has_many_form) + if has_many_form.object.new_record?
-                                      template.content_tag :li do
-                                        template.link_to I18n.t('active_admin.has_many_delete'), "#", :onclick => "$(this).closest('.has_many_fields').remove(); return false;", :class => "button"
-                                      end
-                                    else
-                                    end
+        # @see https://github.com/justinfrench/formtastic/blob/2.2.1/lib/formtastic/helpers/inputs_helper.rb#L373
+        contents = if block.arity == 1  # for backwards compatibility with REE & Ruby 1.8.x
+          block.call(has_many_form)
+        else
+          index = parent_child_index(options[:parent]) if options[:parent]
+          block.call(has_many_form, index)
+        end
+
+        if has_many_form.object.new_record?
+          contents += template.content_tag(:li) do
+            template.link_to I18n.t('active_admin.has_many_delete'), "#", :onclick => "$(this).closest('.has_many_fields').remove(); return false;", :class => "button"
+          end
+        end
+
+        contents
       end
 
       content = with_new_form_buffer do
         template.content_tag :div, :class => "has_many #{association}" do
-          form_buffers.last << template.content_tag(:h3, association.to_s.titlecase)
+          form_buffers.last << template.content_tag(:h3, object.class.reflect_on_association(association).klass.model_name.human(:count => 1.1))
           inputs options, &form_block
 
           # Capture the ADD JS
@@ -82,7 +86,7 @@ module ActiveAdmin
           end
 
           js = template.escape_javascript(js)
-          js = template.link_to I18n.t('active_admin.has_many_new', :model => association.to_s.singularize.titlecase), "#", :onclick => "$(this).before('#{js}'.replace(/NEW_RECORD/g, new Date().getTime())); return false;", :class => "button"
+          js = template.link_to I18n.t('active_admin.has_many_new', :model => object.class.reflect_on_association(association).klass.model_name.human), "#", :onclick => "$(this).before('#{js}'.replace(/NEW_RECORD/g, new Date().getTime())); return false;", :class => "button"
 
           form_buffers.last << js.html_safe
         end
@@ -130,23 +134,36 @@ module ActiveAdmin
       "ActiveAdmin::Inputs::#{as.to_s.camelize}Input"
     end
 
-    def input_class(as)
-      @input_classes_cache ||= {}
-      @input_classes_cache[as] ||= begin
+    # prevent exceptions in production environment for better performance
+    def input_class_with_const_defined(as)
+      input_class_name = custom_input_class_name(as)
+
+      if ::Object.const_defined?(input_class_name)
+        input_class_name.constantize
+      elsif ActiveAdmin::Inputs.const_defined?(input_class_name)
+        active_admin_input_class_name(as).constantize 
+      elsif Formtastic::Inputs.const_defined?(input_class_name)
+        standard_input_class_name(as).constantize 
+      else
+        raise Formtastic::UnknownInputError
+      end
+    end
+
+    # use auto-loading in development environment
+    def input_class_by_trying(as)
+      begin
         begin
-          begin
-            custom_input_class_name(as).constantize
-          rescue NameError
-            begin
-              active_admin_input_class_name(as).constantize
-            rescue NameError
-              standard_input_class_name(as).constantize
-            end
-          end
+          custom_input_class_name(as).constantize
         rescue NameError
-          raise Formtastic::UnknownInputError
+          begin
+            active_admin_input_class_name(as).constantize
+          rescue NameError
+            standard_input_class_name(as).constantize
+          end
         end
       end
+    rescue NameError
+      raise Formtastic::UnknownInputError
     end
 
     private