+++ lib/active_admin/views/index_as_table.rb
@@ -131,7 +131,7 @@ module ActiveAdmin
       # methods for quickly displaying items on the index page
       #
       class IndexTableFor < ::ActiveAdmin::Views::TableFor
-        
+
         # Display a column for checkbox
         def selectable_column
           return unless active_admin_config.batch_actions.any?
@@ -140,7 +140,9 @@ module ActiveAdmin
 
         # Display a column for the id
         def id_column
-          column('ID', :sortable => :id){|resource| link_to resource.id, resource_path(resource), :class => "resource_id_link"}
+          column(resource_class.human_attribute_name(resource_class.primary_key), :sortable => resource_class.primary_key) do |resource| 
+            link_to resource.id, resource_path(resource), :class => "resource_id_link"
+          end
         end
 
         # Adds links to View, Edit and Delete
@@ -149,16 +151,16 @@ module ActiveAdmin
             :name => ""
           }.merge(options)
           column options[:name] do |resource|
-			      links = ''.html_safe
+            links = ''.html_safe
             if controller.action_methods.include?('show')
-              links += link_to I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link"
+              links << link_to(I18n.t('active_admin.view'), resource_path(resource), :class => "member_link view_link")
             end
             if controller.action_methods.include?('edit')
-              links += link_to I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link"
+              links << link_to(I18n.t('active_admin.edit'), edit_resource_path(resource), :class => "member_link edit_link")
             end
             if controller.action_methods.include?('destroy')
-              links += link_to I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :confirm => I18n.t('active_admin.delete_confirmation'), :class => "member_link delete_link"
-            end            
+              links << link_to(I18n.t('active_admin.delete'), resource_path(resource), :method => :delete, :data => {:confirm => I18n.t('active_admin.delete_confirmation')}, :class => "member_link delete_link")
+            end
             links
           end
         end