+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 
 * Created new view components (Footer, TitleBar, Header, UtilityNav) to more
   easily customize the views in Active Admin and per namespace. (@gregbell)
+* All CSS is now encapsulated under the `body.active_admin` class. This may
+  change the precedence of styles that you created to override or use in
+  other areas of your application.
+* Dashboards are now implemented as pages. For more details of how to configure
+  a page, checkout http://activeadmin.info/docs/9-custom-pages.html
+* Batch Actions allows you to select entries on index page and perform
+  an action against them.
 
 ### Deprecations
 
@@ -14,6 +21,13 @@
 * ActiveAdmin::Menu and ActiveAdmin::MenuItem API has changed. If you were
   creating custom menu items, the builder syntax has changed to. Menu#add now
   accepts a MenuItem, instead of building the menu item for you.
+* `ActiveAdmin::Dashboards.build` is deprecated in favour of generating a page
+  and using the new `config.root_to` option.
+* Arbre is now a gem on its own.
+
+### Contributors
+
+456 commits (120 Pull Requests) by 72 contributors.
 
 ## 0.4.4